third_party/WebKit/LayoutTests/svg/dom/getscreenctm-use-with-additional-translation.html

<!DOCTYPE html>
<title>getScreenCTM: &lt;use> additional translation not included</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<svg>
  <defs>
    <rect id="r" width="100" height="100"/>
  </defs>
  <use href="#r" x="100" y="50" fill="blue"/>
</svg>
<script>
test(function() {
  var m_use = document.querySelector('use').getScreenCTM();
  var m_svg = document.querySelector('svg').getScreenCTM();
  for (let prop of ['a', 'b', 'c', 'd', 'e', 'f'])
    assert_equals(m_use[prop], m_svg[prop], prop);
});
</script>
